[fixup][table-planner][table-runtime] Move aggregate and table functions with runtime logic in runtime
