+++ flink-table/flink-table-planner/src/test/resources/org/apache/flink/table/planner/plan/batch/table/SetOperatorsTest.xml
@@ -34,7 +34,7 @@ Calc(select=[EXPR$0 AS a, b, EXPR$1 AS c])
    +- Exchange(distribution=[hash[b]])
       +- LocalHashAggregate(groupBy=[b], select=[b, Partial_SUM(a) AS sum$0, Partial_COUNT(c) AS count$1])
          +- Calc(select=[a0 AS a, b0 AS b, c0 AS c])
-            +- Correlate(invocation=[org$apache$flink$table$planner$functions$tablefunctions$ReplicateRows$211d2b2fa33c76ac8407a4e802c302c0($0, $1, $2, $3)], correlate=[table(ReplicateRows(sum_vcol_marker,a,b,c))], select=[sum_vcol_marker,a,b,c,a0,b0,c0], rowType=[RecordType(BIGINT sum_vcol_marker, INTEGER a, BIGINT b, VARCHAR(2147483647) c, INTEGER a0, BIGINT b0, VARCHAR(2147483647) c0)], joinType=[INNER])
+            +- Correlate(invocation=[org$apache$flink$table$runtime$functions$table$ReplicateRows$95afb438fe6172b6c86dc9db12c8a5de($0, $1, $2, $3)], correlate=[table(ReplicateRows(sum_vcol_marker,a,b,c))], select=[sum_vcol_marker,a,b,c,a0,b0,c0], rowType=[RecordType(BIGINT sum_vcol_marker, INTEGER a, BIGINT b, VARCHAR(2147483647) c, INTEGER a0, BIGINT b0, VARCHAR(2147483647) c0)], joinType=[INNER])
                +- Calc(select=[sum_vcol_marker, a, b, c], where=[(sum_vcol_marker > 0)])
                   +- HashAggregate(isMerge=[true], groupBy=[a, b, c], select=[a, b, c, Final_SUM(sum$0) AS sum_vcol_marker])
                      +- Exchange(distribution=[hash[a, b, c]])
@@ -124,7 +124,7 @@ LogicalProject(b=[$1], c=[$2])
     <Resource name="optimized exec plan">
       <![CDATA[
 Calc(select=[b0 AS b, c0 AS c])
-+- Correlate(invocation=[org$apache$flink$table$planner$functions$tablefunctions$ReplicateRows$4e806505a5bed6f7a784869a97856e7f($0, $1, $2)], correlate=[table(ReplicateRows(sum_vcol_marker,b,c))], select=[sum_vcol_marker,b,c,b0,c0], rowType=[RecordType(BIGINT sum_vcol_marker, BIGINT b, VARCHAR(2147483647) c, BIGINT b0, VARCHAR(2147483647) c0)], joinType=[INNER])
++- Correlate(invocation=[org$apache$flink$table$runtime$functions$table$ReplicateRows$26c9cc6e2d786afbeb35715c89f9e094($0, $1, $2)], correlate=[table(ReplicateRows(sum_vcol_marker,b,c))], select=[sum_vcol_marker,b,c,b0,c0], rowType=[RecordType(BIGINT sum_vcol_marker, BIGINT b, VARCHAR(2147483647) c, BIGINT b0, VARCHAR(2147483647) c0)], joinType=[INNER])
    +- Calc(select=[sum_vcol_marker, b, c], where=[(sum_vcol_marker > 0)])
       +- HashAggregate(isMerge=[true], groupBy=[b, c], select=[b, c, Final_SUM(sum$0) AS sum_vcol_marker])
          +- Exchange(distribution=[hash[b, c]])
