+++ flink-table/flink-table-planner/src/test/resources/org/apache/flink/table/planner/plan/stream/sql/SetOperatorsTest.xml
@@ -59,7 +59,7 @@ LogicalIntersect(all=[true])
     <Resource name="optimized exec plan">
       <![CDATA[
 Calc(select=[c0 AS c])
-+- Correlate(invocation=[org$apache$flink$table$planner$functions$tablefunctions$ReplicateRows$a5a5fb35ffda8b52251be55e3ceaee69($0, $1)], correlate=[table(ReplicateRows($f0,c))], select=[$f0,c,c0], rowType=[RecordType(BIGINT $f0, VARCHAR(2147483647) c, VARCHAR(2147483647) c0)], joinType=[INNER])
++- Correlate(invocation=[org$apache$flink$table$runtime$functions$table$ReplicateRows$cb0cc50dfc1a7a346db5f9f547bf8342($0, $1)], correlate=[table(ReplicateRows($f0,c))], select=[$f0,c,c0], rowType=[RecordType(BIGINT $f0, VARCHAR(2147483647) c, VARCHAR(2147483647) c0)], joinType=[INNER])
    +- Calc(select=[IF((vcol_left_cnt > vcol_right_cnt), vcol_right_cnt, vcol_left_cnt) AS $f0, c], where=[((vcol_left_cnt >= 1) AND (vcol_right_cnt >= 1))])
       +- GroupAggregate(groupBy=[c], select=[c, COUNT(vcol_left_marker) AS vcol_left_cnt, COUNT(vcol_right_marker) AS vcol_right_cnt])
          +- Exchange(distribution=[hash[c]])
@@ -183,7 +183,7 @@ LogicalMinus(all=[true])
     <Resource name="optimized exec plan">
       <![CDATA[
 Calc(select=[c0 AS c])
-+- Correlate(invocation=[org$apache$flink$table$planner$functions$tablefunctions$ReplicateRows$a5a5fb35ffda8b52251be55e3ceaee69($0, $1)], correlate=[table(ReplicateRows(sum_vcol_marker,c))], select=[sum_vcol_marker,c,c0], rowType=[RecordType(BIGINT sum_vcol_marker, VARCHAR(2147483647) c, VARCHAR(2147483647) c0)], joinType=[INNER])
++- Correlate(invocation=[org$apache$flink$table$runtime$functions$table$ReplicateRows$cb0cc50dfc1a7a346db5f9f547bf8342($0, $1)], correlate=[table(ReplicateRows(sum_vcol_marker,c))], select=[sum_vcol_marker,c,c0], rowType=[RecordType(BIGINT sum_vcol_marker, VARCHAR(2147483647) c, VARCHAR(2147483647) c0)], joinType=[INNER])
    +- Calc(select=[sum_vcol_marker, c], where=[(sum_vcol_marker > 0)])
       +- GroupAggregate(groupBy=[c], select=[c, SUM(vcol_marker) AS sum_vcol_marker])
          +- Exchange(distribution=[hash[c]])
