+++ flink-table/flink-table-runtime/src/main/java/org/apache/flink/table/runtime/functions/aggregate/LastValueAggFunction.java
@@ -16,7 +16,7 @@
  * limitations under the License.
  */
 
-package org.apache.flink.table.planner.functions.aggfunctions;
+package org.apache.flink.table.runtime.functions.aggregate;
 
 import org.apache.flink.annotation.Internal;
 import org.apache.flink.table.api.DataTypes;
@@ -24,7 +24,6 @@ import org.apache.flink.table.data.GenericRowData;
 import org.apache.flink.table.data.RowData;
 import org.apache.flink.table.data.StringData;
 import org.apache.flink.table.data.binary.BinaryStringData;
-import org.apache.flink.table.runtime.functions.aggregate.BuiltInAggregateFunction;
 import org.apache.flink.table.types.DataType;
 import org.apache.flink.table.types.logical.LogicalType;
 
@@ -33,13 +32,13 @@ import java.util.List;
 
 import static org.apache.flink.table.types.utils.DataTypeUtils.toInternalDataType;
 
-/** Built-in FIRST_VALUE aggregate function. */
+/** Built-in LAST_VALUE aggregate function. */
 @Internal
-public final class FirstValueAggFunction<T> extends BuiltInAggregateFunction<T, RowData> {
+public final class LastValueAggFunction<T> extends BuiltInAggregateFunction<T, RowData> {
 
     private transient DataType valueDataType;
 
-    public FirstValueAggFunction(LogicalType valueType) {
+    public LastValueAggFunction(LogicalType valueType) {
         this.valueDataType = toInternalDataType(valueType);
     }
 
@@ -55,8 +54,8 @@ public final class FirstValueAggFunction<T> extends BuiltInAggregateFunction<T, 
     @Override
     public DataType getAccumulatorDataType() {
         return DataTypes.ROW(
-                        DataTypes.FIELD("firstValue", valueDataType.nullable()),
-                        DataTypes.FIELD("firstOrder", DataTypes.BIGINT()))
+                        DataTypes.FIELD("lastValue", valueDataType.nullable()),
+                        DataTypes.FIELD("lastOrder", DataTypes.BIGINT()))
                 .bridgedTo(RowData.class);
     }
 
@@ -78,48 +77,47 @@ public final class FirstValueAggFunction<T> extends BuiltInAggregateFunction<T, 
     public RowData createAccumulator() {
         GenericRowData acc = new GenericRowData(2);
         acc.setField(0, null);
-        acc.setField(1, Long.MAX_VALUE);
+        acc.setField(1, Long.MIN_VALUE);
         return acc;
     }
 
     public void accumulate(RowData rowData, Object value) {
         GenericRowData acc = (GenericRowData) rowData;
-        if (value != null && acc.getLong(1) == Long.MAX_VALUE) {
+        if (value != null) {
             acc.setField(0, value);
-            acc.setField(1, System.currentTimeMillis());
         }
     }
 
     public void accumulate(RowData rowData, Object value, Long order) {
         GenericRowData acc = (GenericRowData) rowData;
-        if (value != null && acc.getLong(1) > order) {
+        if (value != null && acc.getLong(1) < order) {
             acc.setField(0, value);
             acc.setField(1, order);
         }
     }
 
-    public void accumulate(RowData rowData, StringData value) {
+    public void accumulate(GenericRowData acc, StringData value) {
         if (value != null) {
-            accumulate(rowData, (Object) ((BinaryStringData) value).copy());
+            accumulate(acc, (Object) ((BinaryStringData) value).copy());
         }
     }
 
-    public void accumulate(RowData rowData, StringData value, Long order) {
+    public void accumulate(GenericRowData acc, StringData value, Long order) {
         if (value != null) {
-            accumulate(rowData, (Object) ((BinaryStringData) value).copy(), order);
+            accumulate(acc, (Object) ((BinaryStringData) value).copy(), order);
         }
     }
 
     public void resetAccumulator(RowData rowData) {
         GenericRowData acc = (GenericRowData) rowData;
         acc.setField(0, null);
-        acc.setField(1, Long.MAX_VALUE);
+        acc.setField(1, Long.MIN_VALUE);
     }
 
     @SuppressWarnings("unchecked")
     @Override
-    public T getValue(RowData acc) {
-        GenericRowData genericAcc = (GenericRowData) acc;
-        return (T) genericAcc.getField(0);
+    public T getValue(RowData rowData) {
+        GenericRowData acc = (GenericRowData) rowData;
+        return (T) acc.getField(0);
     }
 }
