+++ flink-table/flink-table-runtime/src/main/java/org/apache/flink/table/runtime/functions/table/ReplicateRows.java
@@ -16,7 +16,7 @@
  * limitations under the License.
  */
 
-package org.apache.flink.table.planner.functions.tablefunctions;
+package org.apache.flink.table.runtime.functions.table;
 
 import org.apache.flink.api.common.typeinfo.TypeInformation;
 import org.apache.flink.api.common.typeinfo.Types;
